+++ Core/reader.h
@@ -145,10 +145,10 @@ struct Style : Chunk {
 
 // TODO: pos struct and inherit from it?
 struct elementPosition {
-    uint16_t x;
-    uint16_t y;
-    uint16_t w;
-    uint16_t h;
+    uint8_t x;
+    uint8_t y;
+    uint8_t w;
+    uint8_t h;
     uint16_t styleID;
     uint8_t inf;
     struct infiniteElementPosition* infPos;
